+++ src/main/java/com/fhsh/daitda/order/application/command/OrderItemCommand.java
@@ -1,8 +1,8 @@
-package com.fhsh.daitda.order.domain;
+package com.fhsh.daitda.order.application.command;
 
 import java.util.UUID;
 
-public record OrderItemInfo(
+public record OrderItemCommand(
 	UUID productId,
 	String productName,
 	Integer quantity
