+++ src/main/java/com/fhsh/daitda/order/domain/entity/OrderItem.java
@@ -1,6 +1,8 @@
-package com.fhsh.daitda.order.domain;
+package com.fhsh.daitda.order.domain.entity;
 
 import java.util.UUID;
+
+import com.fhsh.daitda.order.domain.entity.OrderProduct;
 
 import jakarta.persistence.Embedded;
 import jakarta.persistence.Entity;
