+++ src/main/java/com/fhsh/daitda/order/domain/enums/OrderStatus.java
@@ -1,4 +1,4 @@
-package com.fhsh.daitda.order.domain;
+package com.fhsh.daitda.order.domain.enums;
 
 public enum OrderStatus {
 	CREATED, COMPLETED, FAILED, CANCELLED
