+++ src/main/java/com/fhsh/daitda/order/domain/repository/OrderRepository.java
@@ -1,7 +1,9 @@
-package com.fhsh.daitda.order.domain;
+package com.fhsh.daitda.order.domain.repository;
 
 import java.util.Optional;
 import java.util.UUID;
+
+import com.fhsh.daitda.order.domain.entity.Order;
 
 public interface OrderRepository {
 	public Order save(Order order);
