+++ src/main/java/com/fhsh/daitda/order/infrastructure/infrastructure/repository/OrderJpaRepository.java
@@ -1,10 +1,10 @@
-package com.fhsh.daitda.order.infrastructure;
+package com.fhsh.daitda.order.infrastructure.infrastructure.repository;
 
 import java.util.UUID;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.fhsh.daitda.order.domain.Order;
+import com.fhsh.daitda.order.domain.entity.Order;
 
 public interface OrderJpaRepository extends JpaRepository<Order, UUID> {
 }
